[Fix] #54 - 위시리스트 추가 DTO 분리
