+++ src/main/java/Fodong/serverdong/domain/wishlist/dto/request/RequestWishlistDto.java
@@ -1,12 +1,7 @@
 package Fodong.serverdong.domain.wishlist.dto.request;
 
-import Fodong.serverdong.domain.category.Category;
-import Fodong.serverdong.domain.member.Member;
-import Fodong.serverdong.domain.restaurant.Restaurant;
-import Fodong.serverdong.domain.wishlist.Wishlist;
 import lombok.*;
 
-import java.time.LocalDateTime;
 @Getter
 @Setter
 @AllArgsConstructor
@@ -14,12 +9,4 @@ import java.time.LocalDateTime;
 public class RequestWishlistDto {
     private Long restaurantId;
 
-    public static Wishlist toEntity(Member member, Restaurant restaurant, Category category) {
-        return Wishlist.builder()
-                .restaurant(restaurant)
-                .category(category)
-                .member(member)
-                .createdDate(LocalDateTime.now())
-                .build();
-    }
 }
